requirements.txt
absl-py==2.1.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
cycler==0.12.1
daphne==4.1.2
decorator==5.1.1
Django==5.0.6
django-debug-toolbar==4.4.2
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
exceptiongroup==1.2.1
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.53.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
ipython==8.25.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
keras==3.3.3
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
msgpack==1.0.8
namex==0.0.8
numba==0.60.0
numpy==1.26.4
openapi-codec==1.3.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
prompt_toolkit==3.0.47
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sounddevice==0.5.0
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.5.0
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-intel==2.16.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
traitlets==5.14.3
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
zope.interface==6.4.post2
